Added callframes with the best efficiency I can think of right now

+++ src/compiler/chunk.rs
@@ -22,12 +22,12 @@ pub enum OpCode {
     OpEqual,
 
     // Primary
-    OpConstant(u8),
-    OpDefineGlobal(u8),
-    OpGetGlobal(u8),
-    OpSetGlobal(u8),
-    OpGetLocal(u8),
-    OpSetLocal(u8),
+    OpConstant(usize),
+    OpDefineGlobal(usize),
+    OpGetGlobal(usize),
+    OpSetGlobal(usize),
+    OpGetLocal(usize),
+    OpSetLocal(usize),
 
     // Unary
     OpNegate,
@@ -73,7 +73,7 @@ impl Chunk {
         self.values.len() - 1
     }
 
-    pub fn get_const(&self, idx: u8) -> Value {
+    pub fn get_const(&self, idx: usize) -> Value {
         self.values[idx as usize].clone()
     }
 }
